requirements.txt
cohere==5.17.0
pinecone==7.3.0
protobuf==6.32.0
python-dotenv==1.1.1
streamlit==1.49.1
tiktoken==0.11.0
google-generativeai==0.8.5
PyPDF2==3.0.1
langchain==0.3.27
langchain-community==0.3.29
langchain-pinecone==0.2.11
langchain-google-genai==2.1.10
langchain-cohere==0.4.5
pypdf==6.0.0
nest_asyncio==1.6.0